+++ tests/Gedmo/Translatable/PersonalTranslationTest.php
@@ -12,6 +12,8 @@ declare(strict_types=1);
 namespace Gedmo\Tests\Translatable;
 
 use Doctrine\Common\EventManager;
+use Doctrine\DBAL\Logging\Middleware;
+use Doctrine\DBAL\ParameterType;
 use Doctrine\ORM\Query;
 use Gedmo\Tests\Tool\BaseTestCaseORM;
 use Gedmo\Tests\Translatable\Fixture\Personal\Article;
@@ -66,12 +68,36 @@ final class PersonalTranslationTest extends BaseTestCaseORM
         $this->populate();
         $this->translatableListener->setTranslatableLocale('lt');
 
-        $this->startQueryLog();
+        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
+        if (class_exists(Middleware::class)) {
+            $this->queryLogger
+                ->expects(static::exactly(2))
+                ->method('debug')
+                ->withConsecutive(
+                    ['Executing statement: {sql} (parameters: {params}, types: {types})', [
+                        'sql' => 'SELECT t0.id AS id_1, t0.title AS title_2 FROM Article t0 WHERE t0.id = ?',
+                        'params' => [1 => 1],
+                        'types' => [1 => ParameterType::INTEGER],
+                    ]],
+                    ['Executing statement: {sql} (parameters: {params}, types: {types})', [
+                        'sql' => 'SELECT t0.id AS id_1, t0.locale AS locale_2, t0.field AS field_3, t0.content AS content_4, t0.object_id AS object_id_5 FROM article_translations t0 WHERE t0.object_id = ?',
+                        'params' => [1 => 1],
+                        'types' => [1 => ParameterType::INTEGER],
+                    ]]
+                );
+        } else {
+            $this->startQueryLog();
+        }
+
         $article = $this->em->find(self::ARTICLE, ['id' => 1]);
 
-        $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
-        static::assertCount(2, $sqlQueriesExecuted);
-        static::assertSame('SELECT t0.id AS id_1, t0.locale AS locale_2, t0.field AS field_3, t0.content AS content_4, t0.object_id AS object_id_5 FROM article_translations t0 WHERE t0.object_id = 1', $sqlQueriesExecuted[1]);
+        // TODO: Remove the "if" block when dropping support of doctrine/dbal 2.
+        if (!class_exists(Middleware::class)) {
+            $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
+            static::assertCount(2, $sqlQueriesExecuted);
+            static::assertSame('SELECT t0.id AS id_1, t0.locale AS locale_2, t0.field AS field_3, t0.content AS content_4, t0.object_id AS object_id_5 FROM article_translations t0 WHERE t0.object_id = 1', $sqlQueriesExecuted[1]);
+        }
+
         static::assertSame('lt', $article->getTitle());
     }
 
@@ -165,15 +191,42 @@ final class PersonalTranslationTest extends BaseTestCaseORM
         $this->em->persist($article);
         $this->em->flush();
 
-        $this->startQueryLog();
+        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
+        if (class_exists(Middleware::class)) {
+            $this->queryLogger
+                ->expects(static::exactly(3))
+                ->method('debug')
+                ->withConsecutive(
+                    ['Beginning transaction'],
+                    ['Executing statement: {sql} (parameters: {params}, types: {types})', [
+                        'sql' => 'UPDATE article_translations SET content = ? WHERE id = ?',
+                        'params' => [
+                            1 => 'change lt',
+                            2 => 1,
+                        ],
+                        'types' => [
+                            1 => ParameterType::STRING,
+                            2 => ParameterType::INTEGER,
+                        ],
+                    ]],
+                    ['Committing transaction']
+                );
+        } else {
+            $this->startQueryLog();
+        }
+
         $this->translatableListener->setTranslatableLocale('lt');
         $article->setTitle('change lt');
 
         $this->em->persist($article);
         $this->em->flush();
-        $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
-        static::assertCount(3, $sqlQueriesExecuted); // one update, transaction start - commit
-        static::assertSame("UPDATE article_translations SET content = 'change lt' WHERE id = 1", $sqlQueriesExecuted[1]);
+
+        // TODO: Remove the "if" block when dropping support of doctrine/dbal 2.
+        if (!class_exists(Middleware::class)) {
+            $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
+            static::assertCount(3, $sqlQueriesExecuted); // one update, transaction start - commit
+            static::assertSame("UPDATE article_translations SET content = 'change lt' WHERE id = 1", $sqlQueriesExecuted[1]);
+        }
     }
 
     public function testShouldBeAbleToUseTranslationQueryHint(): void
@@ -186,14 +239,31 @@ final class PersonalTranslationTest extends BaseTestCaseORM
             ->setHint(TranslatableListener::HINT_TRANSLATABLE_LOCALE, 'lt')
         ;
 
-        $this->startQueryLog();
+        // TODO: Remove the "if" check and "else" body when dropping support of doctrine/dbal 2.
+        if (class_exists(Middleware::class)) {
+            $this->queryLogger
+                ->expects(static::exactly(1))
+                ->method('debug')
+                ->withConsecutive(
+                    ['Executing query: {sql}', [
+                        'sql' => "SELECT CAST(t1_.content AS VARCHAR(128)) AS title_0 FROM Article a0_ LEFT JOIN article_translations t1_ ON t1_.locale = 'lt' AND t1_.field = 'title' AND t1_.object_id = a0_.id",
+                    ]]
+                );
+        } else {
+            $this->startQueryLog();
+        }
+
         $result = $query->getArrayResult();
 
         static::assertCount(1, $result);
         static::assertSame('lt', $result[0]['title']);
-        $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
-        static::assertCount(1, $sqlQueriesExecuted);
-        static::assertSame("SELECT CAST(t1_.content AS VARCHAR(128)) AS title_0 FROM Article a0_ LEFT JOIN article_translations t1_ ON t1_.locale = 'lt' AND t1_.field = 'title' AND t1_.object_id = a0_.id", $sqlQueriesExecuted[0]);
+
+        // TODO: Remove the "if" block when dropping support of doctrine/dbal 2.
+        if (!class_exists(Middleware::class)) {
+            $sqlQueriesExecuted = $this->queryAnalyzer->getExecutedQueries();
+            static::assertCount(1, $sqlQueriesExecuted);
+            static::assertSame("SELECT CAST(t1_.content AS VARCHAR(128)) AS title_0 FROM Article a0_ LEFT JOIN article_translations t1_ ON t1_.locale = 'lt' AND t1_.field = 'title' AND t1_.object_id = a0_.id", $sqlQueriesExecuted[0]);
+        }
     }
 
     protected function getUsedEntityFixtures(): array